First real chunk of code
Old code has been recovered without models and anything indicating if
there is anything else missing.
I will try to rebuild it in C# with SDL, probably the 2.0 and tag it. It
will be a refactor of v1.0
After it, I will modify the necessary stuff to make this code possible to
be proud of. And to be attached into a pipeline. And moddable and
modifiable.